se empezó la modularizacion del tp, sin testear
